lattice file?

+++ crf/lattice/lite/test.py
@@ -2,11 +2,11 @@ import torch
 #import lattice
 from torch.utils.cpp_extension import load
 lattice = load(name="lattice",sources=["lattice.cpp"])
-ref = torch.rand(1000000,3).double()*1000
-src = torch.rand(1000000,6).double()
+ref = torch.rand(1000000,3).float()*1000
+src = torch.rand(1000000,6).float()
 #out = torch.zeros(1000000,16)
 #print(lattice.testfunc(torch.ones((2,3))))
-src[:,-1] = torch.ones(1000000).double()
+src[:,-1] = torch.ones(1000000).float()
 print(ref)
 print(src)
 print(lattice.filter(src,ref))
